backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.0

# Environment variables (Week 1)
python-dotenv==1.0.0

# Database migrations (Week 2)
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Testing (Week 7)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pytest-cov==4.1.0  # Optional: for coverage reports

